[fix]マイグレーションファイルの修正
reservationテーブルにevents,usersの外部参照条件を追記

+++ backend/database/migrations/2025_02_15_053654_create_reservations_table.php
@@ -12,8 +12,8 @@ return new class () extends Migration {
     {
         Schema::create('reservations', function (Blueprint $table) {
             $table->bigIncrements("id");
-            $table->integer("user_id");
-            $table->integer("event_id");
+            $table->unsignedBigInteger("user_id");
+            $table->unsignedBigInteger("event_id");
             $table->integer("quantity");
             $table->tinyInteger("status");
             $table->timestamps();
